update api account
