+++ src/main/java/fall24/swp391/KoiOrderingSystem/model/response/GoogleResponse.java
@@ -1,0 +1,8 @@
+package fall24.swp391.KoiOrderingSystem.model.response;
+
+import lombok.Data;
+
+@Data
+public class GoogleResponse {
+    String token;
+}
