--- src/main/java/fall24/swp391/KoiOrderingSystem/repo/IUserRepository.java
@@ -1,9 +1,0 @@
-package fall24.swp391.KoiOrderingSystem.repo;
-
-import fall24.swp391.KoiOrderingSystem.pojo.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface IUserRepository extends JpaRepository<User,Long> {
-}
